Implement 4.2 and 4.3 C++. Update 4.2 and 4.3 Python.

+++ cpp_solutions/chapter_01_arrays_and_strings/problem_01_02_isPermutation.cpp
@@ -40,10 +40,10 @@ bool chapter_01::isPermutation(const std::string& s1, const std::string& s2){
     }
     std::bitset<128> asciiTable;
     for (char letter : s1){
-        asciiTable.flip((int) letter);
+        asciiTable.flip(static_cast<size_t>(letter));
     }
     for (char letter : s2){
-        asciiTable.flip((int) letter);
+        asciiTable.flip(static_cast<size_t>(letter));
     }
     return asciiTable.none();
 };
